+++ src/App.js
@@ -36,6 +36,7 @@ import Error from './pages/Error';
 
 import {
     ProductsProvider,
+    ProductsConsumer,
     useProductsContext,
 } from './utils/context/ProductsContext';
